journeyman/backend-python/requirements.txt
cryptography>=41.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
APScheduler>=3.10.0
